# docs/swagger.yaml
definitions:
...
  database.Order:
    properties:
      id:
        type: integer
      status:
        type: string
      total_price:
        type: integer
      user_id:
        type: integer
    required:
    - status
    - total_price
    - user_id
    type: object
...
  gin.H:
    additionalProperties: {}
    type: object
...
info:
  contact: {}
  description: REST API for a bookstore with books, orders, and users
  title: Bookstore API
  version: "2.0"
paths:
...
  /api/v1/orders:
    get:
      consumes:
      - application/json
      description: gets a page of orders
      parameters:
      - description: page number to request
        in: query
        name: page
        type: integer
      - description: max number of orders to return per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: successfully got a page of orders
          schema:
            items:
              $ref: '#/definitions/database.Order'
            type: array
        "403":
          description: wrong role
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: error getting a page
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - CookieAuth: []
      summary: gets a page of orders
      tags:
      - order
    post:
      consumes:
      - application/json
      description: creates an order
      parameters:
      - description: new order to add to db
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/database.Order'
      produces:
      - application/json
      responses:
        "201":
          description: successfully created an order
          schema:
            $ref: '#/definitions/database.Order'
        "400":
          description: error binding JSON
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: wrong role
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: error creating order
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - CookieAuth: []
      summary: creates an order
      tags:
      - order
  /api/v1/orders/:id:
    delete:
      consumes:
      - application/json
      description: delete an order by id
      parameters:
      - description: id of order to delete
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: successfully deleted
        "400":
          description: invalid id
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: wrong role
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: error deleting order
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - CookieAuth: []
      summary: delete order
      tags:
      - order
    get:
      consumes:
      - application/json
      description: get one order by id
      parameters:
      - description: id of order to get
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: successfully got an order
          schema:
            $ref: '#/definitions/database.Order'
        "400":
          description: invalid order id
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: wrong role/unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: order not found with this id
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: error getting order
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - CookieAuth: []
      summary: get one order
      tags:
      - order
    put:
      consumes:
      - application/json
      description: update an order by id
      parameters:
      - description: id of order to update
        in: query
        name: id
        required: true
        type: integer
      - description: updated order data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/database.Order'
      produces:
      - application/json
      responses:
        "200":
          description: successfully updated a order
          schema:
            $ref: '#/definitions/database.Order'
        "400":
          description: error binding JSON
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: wrong role/unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: order to update not found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: failed to update order
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - CookieAuth: []
      summary: update an order
      tags:
      - order
...
  /api/v2/orders/all:
    get:
      consumes:
      - application/json
      description: gets all orders
      produces:
      - application/json
      responses:
        "200":
          description: successfully got all Orders
          schema:
            items:
              $ref: '#/definitions/database.Order'
            type: array
        "403":
          description: wrong role
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: error getting all Orders
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - CookieAuth: []
      summary: gets all orders
      tags:
      - order
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: auth_token
    type: apiKey
swagger: "2.0"